+++ app/src/main/java/yukihane/dq10don/communication_game/dto/farm/openalltresurebox/SuccessList.java
@@ -30,7 +30,7 @@ public class SuccessList {
     @JsonProperty("type")
     private String type;
     @JsonProperty("treasureboxTicket")
-    private Integer treasureboxTicket;
+    private Long treasureboxTicket;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
@@ -100,7 +100,7 @@ public class SuccessList {
      *     The treasureboxTicket
      */
     @JsonProperty("treasureboxTicket")
-    public Integer getTreasureboxTicket() {
+    public Long getTreasureboxTicket() {
         return treasureboxTicket;
     }
 
@@ -110,7 +110,7 @@ public class SuccessList {
      *     The treasureboxTicket
      */
     @JsonProperty("treasureboxTicket")
-    public void setTreasureboxTicket(Integer treasureboxTicket) {
+    public void setTreasureboxTicket(Long treasureboxTicket) {
         this.treasureboxTicket = treasureboxTicket;
     }
